requirements.txt
Flask==2.0.1
Werkzeug==2.0.0
Flask-WTF==1.0.0
Flask-Uploads==0.2.1
Pillow==8.3.1
